+++ lib/crc.ex
@@ -4,6 +4,9 @@ defmodule CRC do
   for binary data. It uses NIF functions written in C to iterate over
   the given binary calculating the CRC checksum value.
 
+  `CRC.Pure` provides the same API with no NIF, and will replace the NIF in
+  v1.0. See the "Upgrading to 1.0" section of the README.
+
   CRC implementations have been tested against these online calculators to
   validate their correctness to the best of our ability.
 
@@ -45,76 +48,66 @@ defmodule CRC do
 
       `%{extend: :crc_16_ccitt_false, init: 0x1D0F}`
 
+  The input comes first so calls can be written as pipelines:
+
+      read_data() |> CRC.calculate(:crc_16) |> do_something()
+
   You can learn more about CRC calculation here:
   https://www.sunshine2k.de/articles/coding/crc/understanding_crc.html
   """
-  @spec crc(:crc_algorithm.params(), iodata()) :: :crc_algorithm.value()
-  defdelegate crc(params, input), to: :crc
+  @spec calculate(iodata(), :crc_algorithm.params()) :: :crc_algorithm.value()
+  defdelegate calculate(input, params), to: :crc
 
   @doc """
-  Initialize a resource to be used for doing CRC calculations. The returned
-  resource can be used with `crc/2` or `crc_update/2` to calculate CRC checksums.
+  Initialize a resource to be used for a multi-part CRC calculation with
+  `update/2` and `final/1`.
 
-  Resource is created using the same `params` types that are used with `crc/2`:
+  Resource is created using the same `params` types that are used with
+  `calculate/2`:
 
     - atom's for compiled models
     - Map with model values
     - Map to extend a compiled model.
-
-  If used with `crc/2` the returned resource can be re-used multiple times, but
-  using a map or atom for a compiled model will likely be slightly more
-  performant.
-
-  When using with `crc_update/2` a new resource will be returned with every
-  call that should be used to continue the calculation.
   """
-  @spec crc_init(:crc_algorithm.params()) :: :crc_algorithm.resource()
-  defdelegate crc_init(params), to: :crc
+  @spec init(:crc_algorithm.params()) :: :crc_algorithm.resource()
+  defdelegate init(params), to: :crc
 
   @doc """
   Begins or continues a multi-part CRC calculation.
 
-  Takes a `resource` from result of `crc_init/1` or previous `crc_update/2`
+  Takes a `resource` from result of `init/1` or previous `update/2`
   call, and binary `input`, returns a new `resource` to be used to continue or
   finalize the CRC calculation.
   """
-  @spec crc_update(:crc_algorithm.resource(), iodata()) :: :crc_algorithm.resource()
-  defdelegate crc_update(resource, input), to: :crc
+  @spec update(:crc_algorithm.resource(), iodata()) :: :crc_algorithm.resource()
+  defdelegate update(resource, input), to: :crc
 
   @doc """
-  Takes a `resource` result from `crc_update/2` and finalizes the multi-part
+  Takes a `resource` result from `update/2` and finalizes the multi-part
   CRC calculation.
   """
-  @spec crc_final(:crc_algorithm.resource()) :: :crc_algorithm.value()
-  defdelegate crc_final(resource), to: :crc
+  @spec final(:crc_algorithm.resource()) :: :crc_algorithm.value()
+  defdelegate final(resource), to: :crc
 
   @doc """
-  Calculate a CRC checksum for the `input` based on the crc `params` given.
-
-  See `CRC.crc/2` for details on valid `params`.
-
-  This function has the parameter order reversed to allow easier use with pipelines.
-  allowing code to be written like:
-
-  ## Examples
-
-      read_data() |> CRC.calculate(:crc_16) |> do_something()
-
+  Returns the parameters (`width`, `poly`, `init`, `refin`, `refout`,
+  `xorout`, `check`, `residue` and `sick`) of a model or of a resource
+  returned by `init/1`.
   """
-  @spec calculate(iodata(), :crc_algorithm.params()) :: :crc_algorithm.value()
-  def calculate(input, params) do
-    :crc.crc(params, input)
-  end
+  @spec info(:crc_algorithm.params() | :crc_algorithm.resource()) :: :crc_algorithm.info()
+  defdelegate info(params_or_resource), to: :crc
+
+  @doc """
+  Returns the residue of a model or of a resource returned by `init/1`.
+  """
+  @spec residue(:crc_algorithm.params() | :crc_algorithm.resource()) :: :crc_algorithm.value()
+  defdelegate residue(params_or_resource), to: :crc
 
   @doc """
   Returns a list of all the compiled CRC models.
   """
   @spec list() :: [{atom, String.t()}]
-  def list() do
-    :crc_nif.crc_list()
-    |> Map.to_list()
-    |> Enum.map(fn {model, map} -> {model, Map.get(map, model)} end)
-  end
+  defdelegate list(), to: :crc
 
   @doc """
   Returns a list of all compiled CRC Models that match the filter given.
@@ -123,17 +116,33 @@ defmodule CRC do
   and description.
   """
   @spec list(binary) :: [{atom, String.t()}]
-  def list(filter) do
-    list()
-    |> Enum.filter(&list_filter(&1, filter))
-  end
+  defdelegate list(filter), to: :crc
 
-  defp list_filter({model_atom, model_name}, filter) do
-    atom_string = Atom.to_string(model_atom)
+  @doc """
+  Calculates an XOR checksum for the given binary
+  """
+  @spec checksum_xor(binary) :: number
+  defdelegate checksum_xor(input), to: :crc
 
-    {:ok, rfilter} = Regex.compile(filter)
-    Regex.match?(rfilter, atom_string) or Regex.match?(rfilter, model_name)
-  end
+  @doc "Deprecated, use `CRC.calculate/2` instead (note the argument order: input, params)."
+  @deprecated "Use CRC.calculate/2 instead (note the argument order: input, params)"
+  @spec crc(:crc_algorithm.params(), iodata()) :: :crc_algorithm.value()
+  def crc(params, input), do: :crc_fast.calc(params, input)
+
+  @doc "Deprecated, use `CRC.init/1` instead."
+  @deprecated "Use CRC.init/1 instead"
+  @spec crc_init(:crc_algorithm.params()) :: :crc_algorithm.resource()
+  def crc_init(params), do: :crc_fast.init(params)
+
+  @doc "Deprecated, use `CRC.update/2` instead."
+  @deprecated "Use CRC.update/2 instead"
+  @spec crc_update(:crc_algorithm.resource(), iodata()) :: :crc_algorithm.resource()
+  def crc_update(resource, input), do: :crc_fast.update(resource, input)
+
+  @doc "Deprecated, use `CRC.final/1` instead."
+  @deprecated "Use CRC.final/1 instead"
+  @spec crc_final(:crc_algorithm.resource()) :: :crc_algorithm.value()
+  def crc_final(resource), do: :crc_fast.final(resource)
 
   use CRC.Legacy
 end